requirements.txt
Flask
joblib
numpy
Pillow
deepface
gunicorn